requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
python-dotenv==1.0.0
Markdown==3.5.1
PyYAML==6.0.1
python-dateutil==2.8.2
gunicorn==21.2.0
